// front/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        FieldInstance: {
            readonly id?: number;
            /**
             * Format: iri-reference
             * @example https://example.com/
             */
            form?: string;
            /**
             * Format: iri-reference
             * @example https://example.com/
             */
            component?: string;
            name?: string;
            label?: string;
            /** @default 0 */
            position: number;
            validationRules?: (string | null)[];
            uiOptions?: (string | null)[];
            /** @default default */
            renderStrategy: string;
            renderOptions?: (string | null)[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}